feat: add user controller
